Improve endpoint updates
It is possible to just use the put events command to update endpoints so we can remove the update endpoint command requirement. This also has the effect of ensuring the endpoint is always in sync with the event records as there is no race condition between the initial update endpoint call and the first event being recorded.

Lastly it also removes an unnecessary HTTP request to update endpoint separately.