Support for providing arrays in attributes and metrics
Both AWS Pinpoint and Elasticsearch support arrays of the required datatype eg string or floats for attributes and metrics respectively. We were previsouly sanitising these to individual numbers and strings.

The benefit of having array values for these means that rather than using attribute keys like `category_NN` where `NN` is the ID we could have an attribute containing all category IDs associated with the event eg. `category_ids: [1,2,3]`.

Elasticsearch queries are unaffected by this change in the shape of the data as it does not differentiate between arrays and single values, only the data type is important.

Fixes #46