Add custom event dispatchers
When we figure out the audience(s) an endpoint belongs to elements on the page such as experience blocks may need to respond to these changes to show personalised content.

Because an endpoint can be updated either directly through `Altis.Analytics.updateEndpoint()` or by passing endpoint data to `Altis.Analytics.record()` the endpoints assigned audiences could change and affect content on the page.

For example, we may fetch the current weather conditions async and update the endpoint once complete.