Make audience settings available client side
This exposes the audience configurations and parses them based on the current global attributes, metrics and endpoint data in order to assign the current page view to an audience.

The audiences are updated any time the endpoint data is updated or `registerAttribute()` or `registerMetric()` is called in order to keep them in sync with any custom global attributes or metrics that are derived async.

If the audiences change an event is triggered that allows external code to hook in and respond to the changes.