Fix audience preview loading
The analytics script can be ready before `DOMContentLoaded` has run and in some cases even afterwards because the script is loaded async.

This meant in some set ups that the preview feature was initialising itself too early. This update switches to using the `Altis.Analytics.onReady()` callback and a `useEffect` call to listen for the first `updateAudiences` event to set the defaults.