Integrate Altis Consent
This adds support for the Altis Consent API and automatically integrates it with the analytics layer. We need to do some more QA regarding the behaviour of XBs and AB tests and how they degrade gracefully. Right now XBs display nothing and AB tests just show the default fallback content.