Fix search attribute value
The intention for the search attribute was to store the search term set earlier in the file. It is being replaced by the URL query string currently. Although we can query against the `qv_s` attribute not all sites use the query string for search URLs or the `s` parameter. The search attribute set earlier in the file is populated form the server side so is more reliable.