saving the traffic percentage for post as an array
When updating a post or page the variant traffic percentage would cause an error. 

The function `get_ab_test_variant_traffic_percentage_for_post` would return an an empty string as defined as an array.

The post meta is now checked first if it's an array. It's returned as an array if it is, else an empty array is returned