Fix request ID loggin behaviour
Pinpoint does not have a specific unique ID per event that is available to kinesis or external destinations so this corrects the behaviour for the endpoint request ID value and adds an `x-amz-request-id` attribute to use as an idempotent value for data exports.

The endpoint request ID should only be updated when endpoint data is updated.