Use ternary instead of && so nothing will output if condition fails
The code

    audience && ( <JSX> )

will render "0" into the template if an audience of 0 is passed in,
when you'd expect it to render nothing. This is because JSX faithfully
renders some falsy values as their literal types. Using a ternary with
"null" as the final condition avoids this issue.

(0 will be the default received if you've set up a post meta key for
an audience with an integer value, hypothetically speaking. We can fix
this in application code by passing in `audience || null` as the prop
when using this component, but adjusting it as described here would
make it behave as expected in all circumstances.)