Use parameterised queries for ClickHouse
`$wpdb->prepare` is less appropriate and less powerful compared to ClickHouse built in parameterised value substitution.

This also normalises and corrects a few things such as date ranges, datetime types and the use of the `blog_id` column rather than the attribute as this is part of the sorting key for the table.